Fix Black/Black2 occasional startup crash
This is an attempt to fix a common issue that prevents the Tracker from running. It relates to parsing information from the .pastlog saveData file for BW/BW2. While I don't know the exact cause, the crash seems to be related to trying to iterate on data from a Pokémon that doesn't exist (constData = nil).